feat(configFactory): remove configFactory
