+++ packages/stryker/src/di/buildChildProcessInjector.ts
@@ -1,21 +1,20 @@
-import { Config } from 'stryker-api/config';
 import { commonTokens, Scope, Injector, OptionsContext, tokens } from 'stryker-api/plugin';
-import { optionsFactory, pluginResolverFactory, loggerFactory } from './factoryMethods';
+import { pluginResolverFactory, loggerFactory } from './factoryMethods';
 import { coreTokens } from '.';
 import { getLogger } from 'log4js';
 import { rootInjector } from 'typed-inject';
+import { StrykerOptions } from 'stryker-api/core';
 
-export function buildChildProcessInjector(config: Config): Injector<OptionsContext> {
+export function buildChildProcessInjector(options: StrykerOptions): Injector<OptionsContext> {
   return rootInjector
-    .provideValue(commonTokens.config, config)
-    .provideFactory(commonTokens.options, optionsFactory)
+    .provideValue(commonTokens.options, options)
     .provideValue(commonTokens.getLogger, getLogger)
     .provideFactory(commonTokens.logger, loggerFactory, Scope.Transient)
     .provideFactory(coreTokens.pluginDescriptors, pluginDescriptorsFactory)
     .provideFactory(commonTokens.pluginResolver, pluginResolverFactory);
 }
 
-function pluginDescriptorsFactory(config: Config): ReadonlyArray<string> {
-  return config.plugins;
+function pluginDescriptorsFactory(options: StrykerOptions): ReadonlyArray<string> {
+  return options.plugins;
 }
-pluginDescriptorsFactory.inject = tokens(commonTokens.config);
+pluginDescriptorsFactory.inject = tokens(commonTokens.options);
